+++ components/LayoutOperationsNav.vue
@@ -109,7 +109,7 @@ export default {
     },
     mainHeight () {
       if (process.client) {
-        return (window && window.innerHeight) - 30 - 90
+        return (window && window.innerHeight) - 90
       }
     },
     addOrUpdateFLow () {