+++ CHANGELOG.md
@@ -8,6 +8,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## 0.19.2 - 2023-03-13
+
+### Changed
+* Right sidebar animation and position
+* Transition to item edit
+* Content height
+
+
 ## 0.19.1 - 2023-03-13
 
 ### Fixed